Fix bug when user does not have ssh key
When user did not setup an ssk key the playboo failed.
Now the playbook only distributes the user ssk key if
it exists. If it exists the public key is copied to
the installer machine and OpenShift virtual machines
by default.
This can be disabled by setting set_user_ssh_key to false.

Change-Id: Ideb80f03c22cce9e4292c41b9402eb87c5595392
Signed-off-by: Shirly Radco <sradco@redhat.com>