build: post ovirt-engine-metrics-1.2
ovirt-engine-metrics-1.1, which was part of oVirt 4.2, used OpenShift
3.9, which is unsupported now. We decided to use 3.11, and adapted the
master branch (used for oVirt 4.3 so far) to it. This was done in a
rather long series of patches that also fixed things, and we decided it
will be very hard to cherry-pick changes from there to 4.2. Instead, we
did the following:

1. Renamed branch ovirt-engine-metrics-4.2 to old-ovirt-engine-metrics-4.2.

2. Branched ovirt-engine-metrics-4.2 from current master.

Now, without this patch, 4.2 and master are identical. Current patch
bumps master (only) to 1.3.0, and going forward, we'll use 1.2.Z for the
4.2 branch, and 1.3.Z for 4.3 (currently master branch, eventually we'll
likely branch 4.3 at some point).

Change-Id: I25c0e23fb0ded1451758fa532a2b2de782707505
Signed-off-by: Shirly Radco <sradco@redhat.com>