Added -vvv to the ansible scripts run command
Updated the commad to run the ansible playbooks
in the READMEs to include -vvv, to run them in verbose mode.

It is recommended to run the ansible playbooks
in verbose mode, for better debugging, in case
of a failure.

Change-Id: Iac51d95f6534655e5eb3b01fa225b25aff81af6b
Bug-Url: https://bugzilla.redhat.com/1753941
Signed-off-by: Shirly Radco <sradco@redhat.com>
(cherry picked from commit afe7c377252759f5a45ac45b1d1d3465c1230a8c)